add utils folder, improve response messages and add search param

+++ backend-nodejs-corujao/src/app/controllers/MovieController.js
@@ -1,50 +1,68 @@
 const Movie = require('../models/Movie');
+const SearchWord = require('../utils/SearchWord');
+const ResponseMessage = require('../utils/ResponseMessage');
+const CheckParams = require('../utils/CheckParams');
 
 //Lista os filmes cinematográficos
 module.exports = {
   async listMovies(req, res) {
-    const page = Number(req.query.page) || 1;
-    const size = Number(req.query.size) || 10;
-    const search = req.query.search || '';
+     //Se os valores não forem informados, considera os da direita como default
+     const page = Number(req.query.page) || 1;
+     const size = Number(req.query.size) || 10;
+     const search = req.query.search || '';
     
     try {
-      const movies = await Movie.find()
-      .skip((page - 1) * size)
-      .limit(size)
-      .populate('genres')
-      .populate('director')
-      .populate('cast');
+      const movies = await Movie.find().skip((page - 1) * size).limit(size)
+        .populate('genres')
+        .populate('director')
+        .populate('cast');
 
-      return res.json(movies);
+      //Filtra pelo search caso seja informado o valor
+      if (search !== '') {
+        let moviesFiltered = [];
+        for (let i = 0; i < movies.length; i++) {
+          if (SearchWord.checkIfStringExists(search, movies[i].title))
+            moviesFiltered.push(movies[i]);
+        }        
+        return ResponseMessage.getResponseMessageOK(moviesFiltered, res);
+      }
+
+      return ResponseMessage.getResponseMessageOK(movies, res);
     }
     catch(ex) {
-      return res.status(500).json({ error: 'Erro durante a listagem - server side' });
+      return ResponseMessage.getResponseErrorServerSide(error, res);
     }    
   },
 
   //Cadastra um novo filme
   async addMovie(req, res) {
-    //const { title, releaseYear, genres, director, cast } = req.body;
+    if(!CheckParams.checkIfValidParamsMovie(req.body)) {
+      return ResponseMessage.getResponseErrorClientSide(400, res);
+    }
     
-    //const movie = await Movie.create({ title, releaseYear, genres, director, cast });
     try {
       const movie = await Movie.create(req.body);
+      //const movie = await Movie.create({ title, releaseYear, genres, director, cast });
       
       await movie.populate('genres')
         .populate('director')
         .populate('cast')
         .execPopulate();
       
-      return res.json(movie);
+      return ResponseMessage.getResponseMessageOK(movie, res);
     }
     catch(ex) {
-      return res.status(500).json({ error: 'Erro durante a criação - server side' });
+      return ResponseMessage.getResponseErrorServerSide(error, res);
     }   
   },
 
-  //Detalhe de filme
+  //Detalhe do filme
   async getMovie(req, res) {
-    const movieId = req.params.movieId;
+    const { movieId } = req.params;
+
+    if(!movieId) {
+      return ResponseMessage.getResponseErrorClientSide(400, res);
+    }
 
     try {
       const movie = await Movie.findById(movieId)
@@ -52,54 +70,55 @@ module.exports = {
         .populate('director')
         .populate('cast');
 
-      return res.json(movie);
+      return ResponseMessage.getResponseMessageOK(movie, res);
     } 
     catch(ex) {
-      return res.status(500).json({ error: 'Erro durante a consulta - server side' });
+      return ResponseMessage.getResponseErrorServerSide(error, res);
     }
-
   },
 
-  //Atualização de filme
+  //Atualização do filme
   async updateMovie(req, res) {
-    const movieId = req.params.movieId;
+    const { movieId } = req.params;
     const options = { new: true }; 
 
-    try {
-      
-      movie = await Movie.findByIdAndUpdate(movieId, req.body, options)
+    if(!movieId || !CheckParams.checkIfValidParamsMovie(req.body)) {
+      return ResponseMessage.getResponseErrorClientSide(400, res);
+    }
+
+    try {      
+      const movie = await Movie.findByIdAndUpdate(movieId, req.body, options)
       .populate('genres')
       .populate('director')
       .populate('cast');
 
-      // if (!movie) {
-      //   return res.status(400).json({ error: 'Parâmetros invalidos - client side' });
-      // }
-
-      return res.json(movie);
+      return ResponseMessage.getResponseMessageOK(movie, res);
     }
     catch(ex) {
-      return res.status(500).json({ error: 'Erro durante a atualização - server side' });
+      return ResponseMessage.getResponseErrorServerSide(error, res);
     }    
   },
 
   //Remove um filme do catálogo
   async removeMovie(req, res) {
-    const movieId = req.params.movieId;
+    const { movieId } = req.params;
 
+    if(!movieId) {
+      return ResponseMessage.getResponseErrorClientSide(400, res);
+    }
+    
     try {
       const result = await Movie.deleteOne({ _id: movieId});
 
       if (result.deletedCount === 0) {
-        return res.status(404).json({message: 'Filme não encontrado'});
+        return ResponseMessage.getResponseErrorClientSide(404, res);
       } 
       else {
-        return res.status(204);
-        //return res.status(204).json({message: 'Filme removido com sucesso'});
+        return ResponseMessage.getResponseMessageOKDelete();
       }
     }
     catch(ex) {
-      return res.status(500).json({ error: 'Erro durante a remoção - server side' });
+      return ResponseMessage.getResponseErrorServerSide(error, res);
     }
   }
 }
